+++ src/Command/MigrateDlhCommand.php
@@ -1,31 +1,38 @@
 <?php
 
+/*
+ * Copyright (c) 2020 Heimrich & Hannot GmbH
+ *
+ * @license LGPL-3.0-or-later
+ */
+
 namespace HeimrichHannot\GoogleMapsBundle\Command;
 
 use Contao\Config;
 use Contao\CoreBundle\Command\AbstractLockedCommand;
-use Contao\CoreBundle\Framework\FrameworkAwareInterface;
-use Contao\CoreBundle\Framework\FrameworkAwareTrait;
+use Contao\CoreBundle\Framework\ContaoFramework;
 use Contao\Database;
 use Contao\Model;
 use Contao\StringUtil;
-use Contao\System;
 use HeimrichHannot\GoogleMapsBundle\DataContainer\GoogleMap;
 use HeimrichHannot\GoogleMapsBundle\DataContainer\Overlay;
 use HeimrichHannot\GoogleMapsBundle\Event\DlhMigrationModifyMapEvent;
 use HeimrichHannot\GoogleMapsBundle\Event\DlhMigrationModifyOverlayEvent;
 use HeimrichHannot\GoogleMapsBundle\Model\GoogleMapModel;
 use HeimrichHannot\GoogleMapsBundle\Model\OverlayModel;
+use HeimrichHannot\UtilsBundle\Model\ModelUtil;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\Console\Style\SymfonyStyle;
-use Symfony\Component\EventDispatcher\EventDispatcher;
+use Symfony\Component\EventDispatcher\EventDispatcherInterface;
 
-class MigrateDlhCommand extends AbstractLockedCommand implements FrameworkAwareInterface
+class MigrateDlhCommand extends AbstractLockedCommand
 {
-    use FrameworkAwareTrait;
-
+    /**
+     * @var bool
+     */
+    protected $dryRun = false;
     /**
      * @var SymfonyStyle
      */
@@ -47,18 +54,39 @@ class MigrateDlhCommand extends AbstractLockedCommand implements FrameworkAwareI
     private $cleanBeforeMigration;
 
     /**
-     * @var EventDispatcher
+     * @var EventDispatcherInterface
      */
     private $dispatcher;
+    /**
+     * @var ContaoFramework
+     */
+    private $framework;
+    /**
+     * @var ModelUtil
+     */
+    private $modelUtil;
+
+    public function __construct(string $projectDir, ContaoFramework $framework, EventDispatcherInterface $eventDispatcher, ModelUtil $modelUtil)
+    {
+        parent::__construct();
+        $this->rootDir = $projectDir;
+        $this->framework = $framework;
+        $this->dispatcher = $eventDispatcher;
+        $this->modelUtil = $modelUtil;
+    }
 
     /**
      * {@inheritdoc}
      */
     protected function configure()
     {
-        $this->setName('huh:google-maps:migrate-dlh')->setDescription('Migrates existing Maps created using delahaye/dlh_googlemaps.');
-        $this->addOption('skip-unsupported-field-warnings', null, InputOption::VALUE_NONE, 'Skip warnings indicating that fields don\'t exist anymore in Google Maps v3.');
-        $this->addOption('clean-before-migration', null, InputOption::VALUE_NONE, 'Deletes ALL entries of tl_google_map and tl_google_map_overlay. Use with Caution.');
+        $this
+            ->setName('huh:google-maps:migrate-dlh')
+            ->setDescription('Migrates existing Maps created using delahaye/dlh_googlemaps.')
+            ->addOption('skip-unsupported-field-warnings', null, InputOption::VALUE_NONE, 'Skip warnings indicating that fields don\'t exist anymore in Google Maps v3.')
+            ->addOption('clean-before-migration', null, InputOption::VALUE_NONE, 'Deletes ALL entries of tl_google_map and tl_google_map_overlay. Use with Caution.')
+            ->addOption('dry-run', null, InputOption::VALUE_NONE, 'Performs a run without making changes to the database.')
+        ;
     }
 
     /**
@@ -66,19 +94,27 @@ class MigrateDlhCommand extends AbstractLockedCommand implements FrameworkAwareI
      */
     protected function executeLocked(InputInterface $input, OutputInterface $output)
     {
-        $this->io      = new SymfonyStyle($input, $output);
-        $this->rootDir = $this->getContainer()->getParameter('kernel.project_dir');
+        $this->io = new SymfonyStyle($input, $output);
+
+        $this->io->title('Migrate dlh_googlemaps');
+
+        if ($input->hasOption('dry-run') && $input->getOption('dry-run')) {
+            $this->dryRun = true;
+            $this->io->note('Dry run enabled.');
+            $this->io->newLine();
+        }
+
         $this->framework->initialize();
 
-        $this->dispatcher = System::getContainer()->get('event_dispatcher');
         $this->skipUnsupportedFieldWarnings = $input->getOption('skip-unsupported-field-warnings');
         $this->cleanBeforeMigration = $input->getOption('clean-before-migration');
 
         // clean
-        if ($this->cleanBeforeMigration && $this->io->ask('CAUTION: You set the parameter "clean-before-migration". This will delete ALL entries of tl_google_map and tl_google_map_overlay. Are you sure? [y,n]') === 'y')
-        {
-            Database::getInstance()->execute('DELETE FROM tl_google_map');
-            Database::getInstance()->execute('DELETE FROM tl_google_map_overlay');
+        if ($this->cleanBeforeMigration && $this->io->confirm('CAUTION: You set the parameter "clean-before-migration". This will delete ALL entries of tl_google_map and tl_google_map_overlay. Are you sure?')) {
+            if (!$this->dryRun) {
+                Database::getInstance()->execute('DELETE FROM tl_google_map');
+                Database::getInstance()->execute('DELETE FROM tl_google_map_overlay');
+            }
 
             $this->io->success('tl_google_map and tl_google_map_overlay cleaned successfully.');
         }
@@ -101,7 +137,9 @@ class MigrateDlhCommand extends AbstractLockedCommand implements FrameworkAwareI
         $globalApiKey = Config::get('dlh_googlemaps_apikey');
 
         if ($globalApiKey) {
-            Config::persist('googlemaps_apiKey', $globalApiKey);
+            if (!$this->dryRun) {
+                Config::persist('googlemaps_apiKey', $globalApiKey);
+            }
             $this->io->success('Successfully migrated api key from localconfig.php');
         } else {
             $this->io->caution('No api key found in localconfig.php');
@@ -109,20 +147,22 @@ class MigrateDlhCommand extends AbstractLockedCommand implements FrameworkAwareI
 
         $globalApiKey = Config::get('googlemaps_apiKey');
 
-        if (null !== ($pages = System::getContainer()->get('huh.utils.model')->findAllModelInstances('tl_page'))) {
+        if (null !== ($pages = $this->modelUtil->findAllModelInstances('tl_page'))) {
             foreach ($pages as $page) {
                 if (!($apiKey = $page->dlh_googlemaps_apikey)) {
                     continue;
                 }
 
                 if ($page->googlemaps_apiKey && $page->overrideGooglemaps_apiKey && $page->googlemaps_apiKey != $apiKey) {
-                    $this->io->caution('An api key has been found in the field "dlh_googlemaps_apikey" in page ID ' . $page->id . ', but it couldn\'t be migrated because a differing api key is already set in the field "googlemaps_apiKey".');
+                    $this->io->caution('An api key has been found in the field "dlh_googlemaps_apikey" in page ID '.$page->id.', but it couldn\'t be migrated because a differing api key is already set in the field "googlemaps_apiKey".');
                 } elseif ($globalApiKey && $apiKey !== $globalApiKey) {
                     $page->overrideGooglemaps_apiKey = true;
-                    $page->googlemaps_apiKey         = $apiKey;
-                    $page->save();
+                    $page->googlemaps_apiKey = $apiKey;
+                    if (!$this->dryRun) {
+                        $page->save();
+                    }
 
-                    $this->io->success('Successfully migrated api key for page ID ' . $page->id);
+                    $this->io->success('Successfully migrated api key for page ID '.$page->id);
                 }
             }
         }
@@ -132,35 +172,35 @@ class MigrateDlhCommand extends AbstractLockedCommand implements FrameworkAwareI
     {
         $this->io->text('Step 2: Migrating existing google maps...');
 
-        if (null !== ($legacyMaps = System::getContainer()->get('huh.utils.model')->findAllModelInstances('tl_dlh_googlemaps'))) {
+        if (null !== ($legacyMaps = $this->modelUtil->findAllModelInstances('tl_dlh_googlemaps'))) {
             foreach ($legacyMaps as $legacyMap) {
-                $this->io->text('Migrating dlh google map ID ' . $legacyMap->id . ' ("' . $legacyMap->title . '") ...');
+                $this->io->text('Migrating dlh google map ID '.$legacyMap->id.' ("'.$legacyMap->title.'") ...');
 
-                $map         = new GoogleMapModel();
-                $map->type   = 'base';
+                $map = new GoogleMapModel();
+                $map->type = 'base';
                 $map->tstamp = $map->dateAdded = time();
 
                 /** @var Database $dbAdapter */
                 $dbAdapter = $this->framework->getAdapter(Database::class);
-                $db        = $dbAdapter->getInstance();
+                $db = $dbAdapter->getInstance();
 
                 $legacyFields = $db->getFieldNames('tl_dlh_googlemaps');
 
                 $skipFields = [
                     'id',
                     'tstamp',
-                    'dateAdded'
+                    'dateAdded',
                 ];
 
                 $fieldsMappings = [
-                    'useMapTypeControl'    => 'addMapTypeControl',
-                    'useZoomControl'       => 'addZoomControl',
-                    'useRotateControl'     => 'addRotateControl',
-                    'usePanControl'        => 'addPanControl',
-                    'useScaleControl'      => 'addScaleControl',
+                    'useMapTypeControl' => 'addMapTypeControl',
+                    'useZoomControl' => 'addZoomControl',
+                    'useRotateControl' => 'addRotateControl',
+                    'usePanControl' => 'addPanControl',
+                    'useScaleControl' => 'addScaleControl',
                     'useStreetViewControl' => 'addStreetViewControl',
-                    'useClusterer'         => 'addClusterer',
-                    'mapTypeId'            => 'mapType'
+                    'useClusterer' => 'addClusterer',
+                    'mapTypeId' => 'mapType',
                 ];
 
                 $fieldsToLower = [
@@ -170,7 +210,7 @@ class MigrateDlhCommand extends AbstractLockedCommand implements FrameworkAwareI
                     'mapTypeControlPos',
                     'zoomControlPos',
                     'rotateControlPos',
-                    'streetViewControlPos'
+                    'streetViewControlPos',
                 ];
 
                 $removedFields = [
@@ -182,42 +222,42 @@ class MigrateDlhCommand extends AbstractLockedCommand implements FrameworkAwareI
                     'zoomControlStyle',
                     'scaleControlPos',
                     'parameter',
-                    'moreParameter'
+                    'moreParameter',
                 ];
 
                 $messageFields = [
-                    'staticMapNoscript' => 'The current google map has a static map set. Please set the width and height manually.'
+                    'staticMapNoscript' => 'The current google map has a static map set. Please set the width and height manually.',
                 ];
 
                 foreach ($legacyFields as $legacyField) {
-                    if (in_array($legacyField, $skipFields)) {
+                    if (\in_array($legacyField, $skipFields)) {
                         continue;
                     }
 
-                    if (in_array($legacyField, $removedFields)) {
+                    if (\in_array($legacyField, $removedFields)) {
                         if ($legacyMap->{$legacyField} && !$this->skipUnsupportedFieldWarnings) {
-                            $this->io->caution('The field "' . $legacyField . '" which is different from NULL in the current google map is not used in Google Maps v3 anymore or not supported by this bundle. Please refer to https://developers.google.com/maps/documentation/javascript for further information.');
+                            $this->io->caution('The field "'.$legacyField.'" which is different from NULL in the current google map is not used in Google Maps v3 anymore or not supported by this bundle. Please refer to https://developers.google.com/maps/documentation/javascript for further information.');
                         }
 
                         continue;
                     }
 
-                    if (in_array($legacyField, array_keys($messageFields))) {
+                    if (\in_array($legacyField, array_keys($messageFields))) {
                         $this->io->caution($messageFields[$legacyField]);
                     }
 
-                    $newField    = $legacyField;
+                    $newField = $legacyField;
                     $legacyValue = $legacyMap->{$legacyField};
 
-                    if (in_array($legacyField, $fieldsToLower)) {
+                    if (\in_array($legacyField, $fieldsToLower)) {
                         $legacyValue = strtolower($legacyValue);
                     }
 
-                    if (in_array($legacyField, array_keys($fieldsMappings))) {
+                    if (\in_array($legacyField, array_keys($fieldsMappings))) {
                         $newField = $fieldsMappings[$legacyField];
                     }
 
-                    if (in_array($legacyField, array_keys($fieldsMappings))) {
+                    if (\in_array($legacyField, array_keys($fieldsMappings))) {
                         $newField = $fieldsMappings[$legacyField];
                     }
 
@@ -233,7 +273,7 @@ class MigrateDlhCommand extends AbstractLockedCommand implements FrameworkAwareI
                     $address = $legacyMap->geocoderAddress;
 
                     if ($legacyMap->geocoderCountry) {
-                        $address .= ', ' . $GLOBALS['TL_LANG']['CNT'][$legacyMap->geocoderCountry];
+                        $address .= ', '.$GLOBALS['TL_LANG']['CNT'][$legacyMap->geocoderCountry];
                     }
 
                     $map->centerAddress = $address;
@@ -243,7 +283,7 @@ class MigrateDlhCommand extends AbstractLockedCommand implements FrameworkAwareI
                     if (strpos($legacyMap->center, ',')) {
                         $coordinates = explode(',', $legacyMap->center);
 
-                        if (is_array($coordinates) && count($coordinates) > 1) {
+                        if (\is_array($coordinates) && \count($coordinates) > 1) {
                             $map->centerLat = $coordinates[0];
                             $map->centerLng = $coordinates[1];
                         }
@@ -253,54 +293,57 @@ class MigrateDlhCommand extends AbstractLockedCommand implements FrameworkAwareI
                 // sizing
                 $mapSize = StringUtil::deserialize($legacyMap->mapSize, true);
 
-                if (count($mapSize) > 2) {
+                if (\count($mapSize) > 2) {
                     $map->sizeMode = \HeimrichHannot\GoogleMapsBundle\DataContainer\GoogleMap::SIZE_MODE_STATIC;
 
                     $map->width = serialize([
                         'value' => preg_replace('/[^\d]/i', '', $mapSize[0]),
-                        'unit'  => 'px'
+                        'unit' => 'px',
                     ]);
 
                     $map->height = serialize([
                         'value' => preg_replace('/[^\d]/i', '', $mapSize[1]),
-                        'unit'  => 'px'
+                        'unit' => 'px',
                     ]);
                 } else {
-                    $map->sizeMode     = \HeimrichHannot\GoogleMapsBundle\DataContainer\GoogleMap::SIZE_MODE_ASPECT_RATIO;
+                    $map->sizeMode = \HeimrichHannot\GoogleMapsBundle\DataContainer\GoogleMap::SIZE_MODE_ASPECT_RATIO;
                     $map->aspectRatioX = 16;
                     $map->aspectRatioY = 9;
                 }
 
+                /* @noinspection PhpParamsInspection */
                 $this->dispatcher->dispatch(DlhMigrationModifyMapEvent::NAME, new DlhMigrationModifyMapEvent(
                     $legacyMap,
                     $map
                 ));
 
-                $map->save();
+                if (!$this->dryRun) {
+                    $map->save();
+                }
 
                 // tl_dlh_googlemaps_elements -> tl_google_map_overlay
                 $this->migrateOverlays($legacyMap, $map);
 
-                $this->io->success('Successfully migrated dlh google map ID ' . $legacyMap->id . ' ("' . $legacyMap->title . '") to google map ID ' . $map->id);
+                $this->io->success('Successfully migrated dlh google map ID '.$legacyMap->id.' ("'.$legacyMap->title.'") to google map ID '.$map->id);
             }
         }
     }
 
     protected function migrateOverlays(Model $legacyMap, GoogleMapModel $map)
     {
-        $this->io->text('Migrating overlays of dlh google map ID ' . $legacyMap->id . ' ("' . $legacyMap->title . '") ...');
+        $this->io->text('Migrating overlays of dlh google map ID '.$legacyMap->id.' ("'.$legacyMap->title.'") ...');
 
-        if (null !== ($legacyOverlays = System::getContainer()->get('huh.utils.model')->findModelInstancesBy('tl_dlh_googlemaps_elements', ['pid=?'], [$legacyMap->id]))) {
+        if (null !== ($legacyOverlays = $this->modelUtil->findModelInstancesBy('tl_dlh_googlemaps_elements', ['pid=?'], [$legacyMap->id]))) {
             foreach ($legacyOverlays as $legacyOverlay) {
-                $this->io->text('Migrating dlh google map overlay ID ' . $legacyOverlay->id . ' ("' . $legacyOverlay->title . '") ...');
+                $this->io->text('Migrating dlh google map overlay ID '.$legacyOverlay->id.' ("'.$legacyOverlay->title.'") ...');
 
-                $overlay         = new OverlayModel();
+                $overlay = new OverlayModel();
                 $overlay->tstamp = $overlay->dateAdded = time();
-                $overlay->pid    = $map->id;
+                $overlay->pid = $map->id;
 
                 /** @var Database $dbAdapter */
                 $dbAdapter = $this->framework->getAdapter(Database::class);
-                $db        = $dbAdapter->getInstance();
+                $db = $dbAdapter->getInstance();
 
                 $legacyFields = $db->getFieldNames('tl_dlh_googlemaps_elements');
 
@@ -308,69 +351,69 @@ class MigrateDlhCommand extends AbstractLockedCommand implements FrameworkAwareI
                     'id',
                     'pid',
                     'tstamp',
-                    'dateAdded'
+                    'dateAdded',
                 ];
 
                 $fieldsMappings = [
-                    'iconSRC'         => 'iconSrc',
-                    'markerAction'    => 'clickEvent',
-                    'useRouting'      => 'addRouting',
-                    'linkTitle'       => 'titleText',
-                    'infoWindow'      => 'infoWindowText'
+                    'iconSRC' => 'iconSrc',
+                    'markerAction' => 'clickEvent',
+                    'useRouting' => 'addRouting',
+                    'linkTitle' => 'titleText',
+                    'infoWindow' => 'infoWindowText',
                 ];
 
                 $fieldsToLower = [
                     'type',
                     'markerType',
-                    'markerAction'
+                    'markerAction',
                 ];
 
                 $removedFields = [
                     'hasShadow',
                     'shadowSRC',
-                    'shadowSize'
+                    'shadowSize',
                 ];
 
                 $messageFields = [
-                    'staticMapNoscript' => 'The current google map has a static map set. Please set the width and height manually.'
+                    'staticMapNoscript' => 'The current google map has a static map set. Please set the width and height manually.',
                 ];
 
                 foreach ($legacyFields as $legacyField) {
-                    if (in_array($legacyField, $skipFields)) {
+                    if (\in_array($legacyField, $skipFields)) {
                         continue;
                     }
 
-                    if (in_array($legacyField, $removedFields)) {
+                    if (\in_array($legacyField, $removedFields)) {
                         if ($legacyOverlay->{$legacyField} && !$this->skipUnsupportedFieldWarnings) {
-                            $this->io->caution('The field "' . $legacyField . '" which is different from NULL in the current google map is not used in Google Maps v3 anymore or not supported by this bundle. Please refer to https://developers.google.com/maps/documentation/javascript for further information.');
+                            $this->io->caution('The field "'.$legacyField.'" which is different from NULL in the current google map is not used in Google Maps v3 anymore or not supported by this bundle. Please refer to https://developers.google.com/maps/documentation/javascript for further information.');
                         }
 
                         continue;
                     }
 
-                    if (in_array($legacyField, array_keys($messageFields))) {
+                    if (\in_array($legacyField, array_keys($messageFields))) {
                         $this->io->caution($messageFields[$legacyField]);
                     }
 
-                    $newField    = $legacyField;
+                    $newField = $legacyField;
                     $legacyValue = $legacyOverlay->{$legacyField};
 
-                    if (in_array($legacyField, $fieldsToLower)) {
+                    if (\in_array($legacyField, $fieldsToLower)) {
                         $legacyValue = strtolower($legacyValue);
                     }
 
-                    if (in_array($legacyField, array_keys($fieldsMappings))) {
+                    if (\in_array($legacyField, array_keys($fieldsMappings))) {
                         $newField = $fieldsMappings[$legacyField];
                     }
 
-                    if (in_array($legacyField, array_keys($fieldsMappings))) {
+                    if (\in_array($legacyField, array_keys($fieldsMappings))) {
                         $newField = $fieldsMappings[$legacyField];
                     }
 
                     $overlay->{$newField} = $legacyValue;
                 }
 
-                if ($overlay->clickEvent == 'none') {
+                if ('none' == $overlay->clickEvent) {
                     $overlay->clickEvent = '';
                 }
 
@@ -378,11 +421,11 @@ class MigrateDlhCommand extends AbstractLockedCommand implements FrameworkAwareI
                     $overlay->titleMode = Overlay::TITLE_MODE_TITLE_FIELD;
                 }
 
-                if ($legacyOverlay->markerAction == 'LINK' && $legacyOverlay->linkTitle) {
+                if ('LINK' == $legacyOverlay->markerAction && $legacyOverlay->linkTitle) {
                     $overlay->titleMode = Overlay::TITLE_MODE_CUSTOM_TEXT;
                 }
 
-                if ($legacyOverlay->markerAction == 'INFO') {
+                if ('INFO' == $legacyOverlay->markerAction) {
                     $overlay->clickEvent = 'infowindow';
                 }
 
@@ -393,17 +436,17 @@ class MigrateDlhCommand extends AbstractLockedCommand implements FrameworkAwareI
                     if (strpos($legacyOverlay->singleCoords, ',')) {
                         $coordinates = explode(',', str_replace(' ', '', $legacyOverlay->singleCoords));
 
-                        if (is_array($coordinates) && count($coordinates) > 1) {
+                        if (\is_array($coordinates) && \count($coordinates) > 1) {
                             $overlay->positioningLat = $coordinates[0];
                             $overlay->positioningLng = $coordinates[1];
                         }
                     }
                 } else {
                     $overlay->positioningMode = Overlay::POSITIONING_MODE_STATIC_ADDRESS;
-                    $address                  = $legacyOverlay->geocoderAddress;
+                    $address = $legacyOverlay->geocoderAddress;
 
                     if ($legacyOverlay->geocoderCountry) {
-                        $address .= ', ' . $GLOBALS['TL_LANG']['CNT'][$legacyOverlay->geocoderCountry];
+                        $address .= ', '.$GLOBALS['TL_LANG']['CNT'][$legacyOverlay->geocoderCountry];
                     }
 
                     $overlay->positioningAddress = $address;
@@ -429,26 +472,27 @@ class MigrateDlhCommand extends AbstractLockedCommand implements FrameworkAwareI
                 // sizing
                 $infoWindowSize = StringUtil::deserialize($legacyOverlay->infoWindowSize, true);
 
-                if (count($infoWindowSize) > 2) {
+                if (\count($infoWindowSize) > 2) {
                     $overlay->infoWindowWidth = serialize([
                         'value' => $infoWindowSize[0],
-                        'unit'  => 'px'
+                        'unit' => 'px',
                     ]);
 
                     $overlay->infoWindowHeight = serialize([
                         'value' => $infoWindowSize[1],
-                        'unit'  => 'px'
+                        'unit' => 'px',
                     ]);
                 }
 
                 // anchor
                 $infoWindowAnchor = StringUtil::deserialize($legacyOverlay->infoWindowAnchor, true);
 
-                if (count($infoWindowAnchor) > 2) {
+                if (\count($infoWindowAnchor) > 2) {
                     $overlay->infoWindowAnchorX = $infoWindowAnchor[0];
                     $overlay->infoWindowAnchorY = $infoWindowAnchor[1];
                 }
 
+                /* @noinspection PhpParamsInspection */
                 $this->dispatcher->dispatch(DlhMigrationModifyOverlayEvent::NAME, new DlhMigrationModifyOverlayEvent(
                     $legacyOverlay,
                     $overlay,
@@ -456,9 +500,11 @@ class MigrateDlhCommand extends AbstractLockedCommand implements FrameworkAwareI
                     $map
                 ));
 
-                $overlay->save();
+                if (!$this->dryRun) {
+                    $overlay->save();
+                }
 
-                $this->io->success('Successfully migrated dlh google map overlay ID ' . $legacyMap->id . ' ("' . $legacyOverlay->title . '") to google map overlay ID ' . $overlay->id);
+                $this->io->success('Successfully migrated dlh google map overlay ID '.$legacyMap->id.' ("'.$legacyOverlay->title.'") to google map overlay ID '.$overlay->id);
             }
         }
     }