+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [1.5.0-DEV] - 2020-06-17
+- added dry-run option to migration command
+- set migration command dependencies by dependency injection
+
+## [1.4.1] - 2020-04-15
+- fixed MapManager return value declaration
+
 ## [1.4.0] - 2019-12-02
 - also consider googlemaps_apiKey config key when checking for api key (MapManager)
 - removed symfony framework dependency